Do some refactoring of route macro
